% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de-pseudobulk.R
\name{filter_pseudobulk_samples}
\alias{filter_pseudobulk_samples}
\title{Filter pseudobulk samples by size}
\usage{
filter_pseudobulk_samples(pb, min_cells = 10L, min_lib_size = NULL)
}
\arguments{
\item{pb}{A pseudobulk \code{SingleCellExperiment}, typically returned by
\code{aggregate_pseudobulk()}.}

\item{min_cells}{Minimum number of cells required.}

\item{min_lib_size}{Minimum library size required.}
}
\value{
A filtered pseudobulk \code{SingleCellExperiment}.
}
\description{
Filter a pseudobulk \code{SingleCellExperiment} using minimum numbers of cells and
library size.
}
