% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de-edger.R
\name{plot_pairwise_de_volcano}
\alias{plot_pairwise_de_volcano}
\title{Plot a volcano plot for a pairwise edgeR contrast}
\usage{
plot_pairwise_de_volcano(
  de_result,
  contrast = NULL,
  fdr_cutoff = 0.05,
  lfc_cutoff = 0,
  top_n_labels = 10L,
  point_size = 1.5,
  point_alpha = 0.75
)
}
\arguments{
\item{de_result}{Result list returned by \code{run_edger_pairwise_de()}.}

\item{contrast}{Name of the contrast to plot. If \code{NULL}, the first available
contrast is used.}

\item{fdr_cutoff}{FDR threshold used to define significant genes.}

\item{lfc_cutoff}{Absolute log-fold-change threshold used to highlight larger
effects.}

\item{top_n_labels}{Number of top genes to label by FDR. Set to \code{0} to omit
labels.}

\item{point_size}{Point size.}

\item{point_alpha}{Point alpha.}
}
\value{
A \code{ggplot2} object.
}
\description{
Visualize one pairwise contrast from \code{\link[=run_edger_pairwise_de]{run_edger_pairwise_de()}} as a volcano
plot using log-fold change and FDR.
}
