% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tata-plot.R
\name{plot_tata_results}
\alias{plot_tata_results}
\title{Plot an overview of TATA results}
\usage{
plot_tata_results(tata_result, dimred = "UMAP", n_probability_cells = 6)
}
\arguments{
\item{tata_result}{Result list returned by \code{run_tata()}.}

\item{dimred}{Reduced dimension to use for the embedding plots.}

\item{n_probability_cells}{Number of representative cells to show in the
terminal-probability barplot.}
}
\value{
A named list of \code{ggplot2} objects.
}
\description{
Create a compact set of standard TATA result plots analogous to a trajectory
overview dashboard.
}
