% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scProkaR-package.R
\docType{package}
\name{SCProkaR-package}
\alias{SCProkaR}
\alias{SCProkaR-package}
\title{SCProkaR: Microbial Single-Cell RNA-Seq Analysis for Bacterial Systems}
\description{
\code{SCProkaR} provides a \code{SingleCellExperiment}-centered toolkit for microbial
single-cell analysis in R. The package combines bacterial QC and integration
workflows with pseudobulk differential expression utilities and Time Aware
Trajectory Analysis (TATA) for time-informed trajectory abstraction.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MonashLiLab/SCProkaR}
  \item Report bugs at \url{https://github.com/MonashLiLab/SCProkaR/issues}
}

}
\author{
\strong{Maintainer}: Jinxin Zhao \email{jinxin.zhao@example.org}

Authors:
\itemize{
  \item Ning Liu \email{ning.liu@adelaide.edu.au}
}

}
\keyword{internal}
